# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "RenovAI API",
  "description": "Agentic AI platform for interior design & renovation: intake, vision, RAG, deterministic pricing, quotes, negotiation, approvals and CRM sync.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/rag/documents": {
   "get": {
    "tags": [
     "rag"
    ],
    "summary": "List Documents",
    "operationId": "list_documents_api_v1_rag_documents_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "doc_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Doc Type"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response List Documents Api V1 Rag Documents Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "rag"
    ],
    "summary": "Upload Document",
    "operationId": "upload_document_api_v1_rag_documents_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_document_api_v1_rag_documents_post"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Upload Document Api V1 Rag Documents Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/rag/documents/{doc_id}": {
   "delete": {
    "tags": [
     "rag"
    ],
    "summary": "Delete Document",
    "operationId": "delete_document_api_v1_rag_documents__doc_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Doc Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/rag/search": {
   "post": {
    "tags": [
     "rag"
    ],
    "summary": "Search",
    "operationId": "search_api_v1_rag_search_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RagSearch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Search Api V1 Rag Search Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/crm/status": {
   "get": {
    "tags": [
     "crm"
    ],
    "summary": "Crm Status",
    "operationId": "crm_status_api_v1_crm_status_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Crm Status Api V1 Crm Status Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/crm/syncs": {
   "get": {
    "tags": [
     "crm"
    ],
    "summary": "List Syncs",
    "operationId": "list_syncs_api_v1_crm_syncs_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response List Syncs Api V1 Crm Syncs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/crm/syncs/{sync_id}/retry": {
   "post": {
    "tags": [
     "crm"
    ],
    "summary": "Retry Sync",
    "operationId": "retry_sync_api_v1_crm_syncs__sync_id__retry_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "sync_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Sync Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Retry Sync Api V1 Crm Syncs  Sync Id  Retry Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/leads": {
   "get": {
    "tags": [
     "crm"
    ],
    "summary": "List Leads",
    "operationId": "list_leads_api_v1_leads_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response List Leads Api V1 Leads Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/leads/{lead_id}": {
   "patch": {
    "tags": [
     "crm"
    ],
    "summary": "Patch Lead",
    "operationId": "patch_lead_api_v1_leads__lead_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lead_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lead Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LeadPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Patch Lead Api V1 Leads  Lead Id  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/tasks": {
   "get": {
    "tags": [
     "crm"
    ],
    "summary": "List Tasks",
    "operationId": "list_tasks_api_v1_tasks_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "default": "open",
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response List Tasks Api V1 Tasks Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/tasks/{task_id}": {
   "patch": {
    "tags": [
     "crm"
    ],
    "summary": "Patch Task",
    "operationId": "patch_task_api_v1_tasks__task_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "task_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Task Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TaskPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Patch Task Api V1 Tasks  Task Id  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/materials": {
   "get": {
    "tags": [
     "pricing"
    ],
    "summary": "List Materials",
    "operationId": "list_materials_api_v1_materials_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "group",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Group"
      }
     },
     {
      "name": "tier",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Tier"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response List Materials Api V1 Materials Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/materials/{sku}": {
   "put": {
    "tags": [
     "pricing"
    ],
    "summary": "Update Material",
    "operationId": "update_material_api_v1_materials__sku__put",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "sku",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Sku"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MaterialUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Update Material Api V1 Materials  Sku  Put"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/pricing/config": {
   "get": {
    "tags": [
     "pricing"
    ],
    "summary": "Pricing Config",
    "operationId": "pricing_config_api_v1_pricing_config_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Pricing Config Api V1 Pricing Config Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_upload_document_api_v1_rag_documents_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     },
     "doc_type": {
      "type": "string",
      "title": "Doc Type",
      "default": "document"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "room_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Room Type"
     },
     "location": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Location"
     },
     "budget_range": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Budget Range"
     },
     "material_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Material Type"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_upload_document_api_v1_rag_documents_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LeadPatch": {
    "properties": {
     "status": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "new",
         "contacted",
         "quoted",
         "won",
         "lost"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Status"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "title": "LeadPatch"
   },
...
   "MaterialUpdate": {
    "properties": {
     "unit_price": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Unit Price"
     },
     "available": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Available"
     },
     "lead_days": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 365.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Lead Days"
     }
    },
    "type": "object",
    "title": "MaterialUpdate"
   },
...
   "RagSearch": {
    "properties": {
     "query": {
      "type": "string",
      "maxLength": 300,
      "minLength": 2,
      "title": "Query"
     },
     "doc_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Doc Type"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "top_k": {
      "type": "integer",
      "maximum": 20.0,
      "minimum": 1.0,
      "title": "Top K",
      "default": 6
     }
    },
    "type": "object",
    "required": [
     "query"
    ],
    "title": "RagSearch"
   },
...
   "TaskPatch": {
    "properties": {
     "status": {
      "type": "string",
      "enum": [
       "open",
       "done"
      ],
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "status"
    ],
    "title": "TaskPatch"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}